Fix cuvs-lucene artifact publication
This commit fixes the publication of cuvs-lucene artifacts to Github.

The Github actions build-logs indicate that while the JAR is built
correctly, it doesn't seem to be uploaded properly as a Github artifact.
It appears to be the result of a minor typo in the artifact's base dir.

This commit fixes the typo, and allows the artifact to be published.

Signed-off-by: MithunR <mithunr@nvidia.com>